begin work on unions; partial progress.
Still looking for a good way to structure this argument parsing code.
It gets complex quickly.  A disambiguation section that runs first
followed by smaller functions for specific styles once distinguished
seems potentially nice.  Worth trying, anyway.

In the course of outlining this, I realized that we really need three
entrypoints to construction, for every type: there should be
a type-level constructor, a representation-level constructor, and
a "do what I mean" constructor.  The latter should probably be the
most syntactically reachable, but all three are important if one should
be able to write code that's completely free of edge cases.

(For example: if you have a union which has keyed representation, and
type names which just happen to match the discriminator strings,
it could be ambiguous when using an object restructuring positional
style whether you meant to engage typed mode or representation mode --
and that might matter if you have other types beneath (or if the
discriminant string is present but maps to a different type, monsterous
though that would be).)