More wip: exploration of the different calling conventions one could find for unions.
(This is a sketch from the last time I had this repo open, and already
looks stale and goofy to me, so we'll be moving on quickly.)

I'm not sure what I was thinking, writing this in the golang docs;
it belongs in real docs, first of all;
and it needs to be based on principles that have consistency
at a broader level than just the union type,
and refer to those.

So that's what'll be up next: docs on the principles.