Prototype(kwargs) builds a dict, is not necessarily Map
When a Prototype is called with kwargs, convert them to a dict, then pass to assembleVal. This is because the result is not necessarily a map (could be a map, struct, or union). Add TODOs for the pieces of missing logic, like Union acting differently when given kwags, and the need for assembleVal to type-check its arguments against the type it is trying to construct.

Other various style cleanups, and tweaks to comments.