Retain compound keys in maps, to fix remaining test
Somewhat of a hack: in order to support maps with compound keys (struct values), retain the original starlark values in the ArgSeq. Given how this feature works, it seems like a bit of special case. Figure out how the rules for assignments to structs work (with regular string-named fields) then come back to this test.